requirements.txt
requests==2.31.0
pyarrow==14.0.2
duckdb==0.9.2
python-dotenv==1.0.0
